Update models structure
