+++ moneybox/clients/currency/client.py
@@ -13,12 +13,16 @@ class ExchangeRatesAPIClient:
         self.api_key = api_key
 
     def get_currencies(self) -> dict[str, str]:
-        response = requests.get(f"{self.CURRENCIES}?app_id={self.api_key}", timeout=10)
+        response = requests.get(
+            f"{self.CURRENCIES}?app_id={self.api_key}", timeout=10
+        )
         response.raise_for_status()
         return response.json()
 
     def get_rates(self):
-        response = requests.get(f"{self.LATEST}?app_id={self.api_key}", timeout=10)
+        response = requests.get(
+            f"{self.LATEST}?app_id={self.api_key}", timeout=10
+        )
         response.raise_for_status()
         resp = Response(**response.json())
         return resp.rates
