+++ moneybox/moneybox/urls.py
@@ -5,6 +5,8 @@ from api.urls import urlpatterns as api_urlpatterns
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))
+    path('api-auth/', include(
+        'rest_framework.urls', namespace='rest_framework')
+         )
 ]
 urlpatterns += api_urlpatterns
