+++ moneybox/wallet/admin/__init__.py
@@ -1,1 +1,21 @@
+from .currency_admin import CurrencyAdmin, CurrencyRateAdmin
+from .expense_admin import ExpenseCategoryAdmin, ExpenseAdmin
+from .group_admin import GroupAdmin
+from .income_admin import IncomeCategoryAdmin, IncomeAdmin
+from .profile_admin import ProfileAdmin
+from .transfer_admin import TransferAdmin
+from .wallet_admin import WalletAdmin
 
+
+__all__ = [
+    'CurrencyAdmin',
+    'CurrencyRateAdmin',
+    'ExpenseAdmin',
+    'ExpenseCategoryAdmin',
+    'GroupAdmin',
+    'IncomeAdmin',
+    'IncomeCategoryAdmin',
+    'ProfileAdmin',
+    'TransferAdmin',
+    'WalletAdmin',
+]
