+++ moneybox/wallet/models/__init__.py
@@ -1,0 +1,20 @@
+from .currency_models import Currency, CurrencyRate
+from .expense_models import Expense, ExpenseCategory
+from .group_models import Group
+from .income_models import Income, IncomeCategory
+from .transfer_models import Transfer
+from .timestamp_models import TimestampMixin
+from .wallet_models import Wallet
+
+__all__ = [
+    'Currency',
+    'CurrencyRate',
+    'Expense',
+    'ExpenseCategory',
+    'Group',
+    'Income',
+    'IncomeCategory',
+    'Transfer',
+    'TimestampMixin',
+    'Wallet',
+]
